Fix q_free by releasing corrective memory
The issue was caused by mixing up the variables "entry" (which can be
modified) and "safe" (which stores correct information for iteration)
in list_for_each_entry_safe. As a result, the wrong variable was being
freed. This is now fixed by ensuring the correct variable is freed
during list traversal.

Change-Id: I2f0dc54a05c26c240e2c6a65aac26a957354dfd1