Implement q_strncmp
Implementing "q_strncmp" function, which works like "strncmp" but
requires list_head type of pointer as parameter. This eliminates the
need to manually use "list_entry" and calculate the n value for
"strncmp".

The main reason for implementing this function was that I noticed that
"q_delete_dup" was difficult to read and contained errors.
Specifically, I forgot to recalculate the n value within the while
loop. I will refactor and provide a detailed explanation of the errors
in "q_delete_dup" in the next commit.

Change-Id: I53156f35f77bc9d7c132be0e4a9c4a94d9d7d6ec