Refactor q_delete_dup for a bug and readability
The original code was not only difficult to understand but also
contained a bug. In the while loop, n was only calculated once, but
since the pointer "to" keeps moving, the string length changes. If "to"
points to a longer string than original "to" and "from", "strncmp" will
fail to compare the entire string, leading to incorrect results. Now,
I use "q_strncmp" function to improve code readability and resolve this
issue.

Change-Id: I752bb45f79eb7f452f2f342100bb159f1f6f7240