Rewrite q_shuffle for randomness quality
The original "q_shuffle" function does not produce a uniform shuffle
when the queue length is large because it uses the modulus operator
with "RAND_MAX". While it can still shuffle the queue, the randomness
quality is low. I have improved the method for generating random numbers
to better handle large queue sizes. However, be aware that if the queue
size is significantly larger than RAND_MAX, the randomness quality may
still be discounted. Nevertheless, this approach provides much better
randomness than the original implementation.

Change-Id: I7e6446d16924440b679a60c393c6b83a5a1529c9